BREAKING CHANGE: API change optical elements
- Renamed the `OpticalElementInterface` class to `BeamTransformation_Interface`
- Renamed the `BaseOpticalElement` class to `BeamTransformation_Base`
- GaussianBeam.transform(...) now takes a reference, instead of a pointer

Renamed the base classes used to implement optical elements to be more
general. For example, "free space propagation" can be represented as a
RTM, but this just corresponds to a shift in the z-axis and is not
really an "optical element".

We also modified the GaussianBeam interface to take a reference to
BeamTransformation_Interface instead of a pointer so that we can pass
instances directly, instead of taking the address of the instance. i.e.

before:

        beam.transform(&lens)

now:

        beam.transform(lens)